refactor: decompose receipt-upload-form.tsx into focused sub-components (#144)
Split 847-line monolithic component into 7 focused modules under
src/components/receipt/: types, custom hook, upload zone, parsing
indicator, line item row, receipt review, and slim orchestrator.
Original file re-exports for backward compatibility.
